+++ virtuous/internal/testtypes/b/user.go
@@ -1,0 +1,5 @@
+package b
+
+type User struct {
+	ID int64
+}